feat: ui와 로직 분리 (FE-65)
